+++ packages/gi-assets-advance/src/components/GremlinQuery/Component.tsx
@@ -1,7 +1,7 @@
 import { useContext, utils } from '@antv/gi-sdk';
 import GremlinEditor from 'ace-gremlin-editor';
-import { Button, Col, InputNumber, notification, Row, Space } from 'antd';
-import React, { useEffect, useState } from 'react';
+import { Button, Col, InputNumber, notification, Row, Space, Tooltip } from 'antd';
+import React, { useEffect, useMemo, useState } from 'react';
 import { useImmer } from 'use-immer';
 import { nanoid } from 'nanoid';
 import './index.less';
@@ -36,6 +36,7 @@ const GremlinQueryPanel: React.FC<IGremlinQueryProps> = ({
   controlledValues,
 }) => {
   const { updateContext, transform, services, updateHistory } = useContext();
+  const constainerRef = React.useRef<null | HTMLDivElement>(null);
 
   const service = utils.getService(services, serviceId);
   const gremlinFromUrl = utils.searchParamOf('gremlin');
@@ -52,7 +53,7 @@ const GremlinQueryPanel: React.FC<IGremlinQueryProps> = ({
     isFullScreen: false,
     modalVisible: false,
     limit: null,
-    timeout: null,
+    timeout: 5000,
   });
 
   const setEditorValue = val => {
@@ -68,6 +69,9 @@ const GremlinQueryPanel: React.FC<IGremlinQueryProps> = ({
 
   const [btnLoading, setBtnLoading] = useState(false);
 
+  const editorValueHasLimit = useMemo(() => editorValue.includes('.limit'), [editorValue]);
+  const editorValueHasTimeout = useMemo(() => editorValue.includes('evaluationTimeout'), [editorValue]);
+
   const handleClickQuery = async () => {
     setBtnLoading(true);
     if (!service) {
@@ -75,10 +79,10 @@ const GremlinQueryPanel: React.FC<IGremlinQueryProps> = ({
     }
 
     let gremlinCode = `${editorValue}`;
-    if (limit) {
+    if (limit && !editorValueHasLimit) {
       gremlinCode = `${gremlinCode}.limit(${limit})`;
     }
-    if (timeout) {
+    if (timeout && !editorValueHasTimeout) {
       gremlinCode = `${gremlinCode.substring(0, 1)}.with('evaluationTimeout', ${timeout})${gremlinCode.substring(1)}`;
     }
 
@@ -145,8 +149,129 @@ const GremlinQueryPanel: React.FC<IGremlinQueryProps> = ({
     }
   }, [controlledValues]);
 
+  const hasConstraint = isShowLimit || isShowTimeout;
+
+  const constraintsHorizontal = useMemo(() => {
+    if (!constainerRef.current?.scrollWidth) return false;
+    if (constainerRef.current?.scrollWidth > 400) {
+      return true;
+    }
+    return false;
+  }, [constainerRef.current?.scrollWidth]);
+
+  const limitConstraint = isShowLimit ? (
+    <div className="gi-gremlin-query-config" style={{ width: constraintsHorizontal ? '50%' : 'unset' }}>
+      <Tooltip
+        title={
+          editorValueHasLimit
+            ? $i18n.get({
+                id: 'advance.components.GremlinQuery.Component.CannotBeModifiedLimitNum',
+                dm: '不可修改，语句中已有 .limit(num) 约束',
+              })
+            : ''
+        }
+      >
+        <div className="gi-gremlin-query-constraint-constraint">
+          <label className="gi-gremlin-query-constraint-label">Limit: </label>
+          <InputNumber
+            min={1}
+            max={Infinity}
+            value={editorValueHasLimit ? undefined : limit}
+            size="small"
+            style={{ width: 'calc(100% - 80px)' }}
+            disabled={editorValueHasLimit}
+            onChange={val =>
+              setState(draft => {
+                draft.limit = val;
+              })
+            }
+          />
+        </div>
+      </Tooltip>
+      {limit && !editorValueHasLimit ? (
+        <div className="gi-gremlin-query-config-tip">
+          {$i18n.get(
+            {
+              id: 'advance.components.GremlinQuery.Component.LimitConfigTip',
+              dm: $i18n.get(
+                {
+                  id: 'advance.components.GremlinQuery.Component.TheStatementAddsASuffix',
+                  dm: '语句将添加后缀: .limt({limit})',
+                },
+                { limit: limit },
+              ),
+            },
+            {
+              value: limit,
+            },
+          )}
+        </div>
+      ) : (
+        ''
+      )}
+    </div>
+  ) : (
+    ''
+  );
+
+  const timeoutConstraint = isShowTimeout ? (
+    <div className="gi-gremlin-query-config" style={{ width: constraintsHorizontal ? '50%' : 'unset' }}>
+      <Tooltip
+        title={
+          editorValueHasTimeout
+            ? $i18n.get({
+                id: 'advance.components.GremlinQuery.Component.CannotBeModifiedWithEvaluationtimeout',
+                dm: "不可修改，语句中已有 .with('evaluationTimeout', timeout) 约束",
+              })
+            : ''
+        }
+      >
+        <div className="gi-gremlin-query-constraint-constraint">
+          <label className="gi-gremlin-query-constraint-label">Timeout(ms): </label>
+          <InputNumber
+            min={500}
+            max={Infinity}
+            step={1000}
+            value={editorValueHasTimeout ? undefined : timeout}
+            size="small"
+            disabled={editorValueHasTimeout}
+            style={{ width: 'calc(100% - 80px)' }}
+            onChange={val =>
+              setState(draft => {
+                draft.timeout = val;
+              })
+            }
+          />
+        </div>
+      </Tooltip>
+      {timeout && !editorValueHasTimeout ? (
+        <div className="gi-gremlin-query-config-tip">
+          {$i18n.get(
+            {
+              id: 'advance.components.GremlinQuery.Component.TimeoutConfigTip',
+              dm: $i18n.get(
+                {
+                  id: 'advance.components.GremlinQuery.Component.StatementToAddConstraintsWith',
+                  dm: "语句将添加约束: .with('evaluationTimeout', {timeout})",
+                },
+                { timeout: timeout },
+              ),
+            },
+            {
+              value: timeout,
+            },
+          )}
+        </div>
+      ) : (
+        ''
+      )}
+    </div>
+  ) : (
+    ''
+  );
+
   return (
-    <div className="gi-gremlin-query " style={{ ...style }}>
+    <div className="gi-gremlin-query " style={{ ...style }} ref={constainerRef}>
       <div style={{ border: '1px solid #f6f6f6' }}>
         <GremlinEditor
           initValue={editorValue}
@@ -156,83 +281,13 @@ const GremlinQueryPanel: React.FC<IGremlinQueryProps> = ({
         />
       </div>
 
-      {isShowLimit ? (
-        <div className="gi-gremlin-query-config">
-          <Row>
-            <Col span={6}>
-              <label>Limit: </label>
-            </Col>
-            <Col span={18}>
-              <InputNumber
-                min={1}
-                max={Infinity}
-                value={limit}
-                size="small"
-                style={{ width: '100%' }}
-                onChange={val =>
-                  setState(draft => {
-                    draft.limit = val;
-                  })
-                }
-              />
-            </Col>
-          </Row>
-          {limit ? (
-            <div className="gi-gremlin-query-config-tip">
-              {$i18n.get(
-                {
-                  id: 'advance.components.GremlinQuery.Component.LimitConfigTip',
-                  dm: `语句将添加后缀: .limt(${limit})`,
-                },
-                {
-                  value: limit,
-                },
-              )}
-            </div>
-          ) : (
-            ''
-          )}
-        </div>
-      ) : (
-        ''
-      )}
-
-      {isShowTimeout ? (
-        <div className="gi-gremlin-query-config">
-          <Row>
-            <Col span={6}>
-              <label>Timeout: </label>
-            </Col>
-            <Col span={18}>
-              <InputNumber
-                min={1}
-                max={Infinity}
-                value={timeout}
-                size="small"
-                style={{ width: '100%' }}
-                onChange={val =>
-                  setState(draft => {
-                    draft.timeout = val;
-                  })
-                }
-              />
-            </Col>
-          </Row>
-          {timeout ? (
-            <div className="gi-gremlin-query-config-tip">
-              {$i18n.get(
-                {
-                  id: 'advance.components.GremlinQuery.Component.TimeoutConfigTip',
-                  dm: `语句将添加约束: .with('evaluationTimeout', ${timeout})`,
-                },
-                {
-                  value: timeout,
-                },
-              )}
-            </div>
-          ) : (
-            ''
-          )}
+      {hasConstraint ? (
+        <div
+          className="gi-gremlin-query-constraints-wrapper"
+          style={{ display: constraintsHorizontal ? 'inline-flex' : 'block' }}
+        >
+          {limitConstraint}
+          {timeoutConstraint}
         </div>
       ) : (
         ''
@@ -259,18 +314,18 @@ const GremlinQueryPanel: React.FC<IGremlinQueryProps> = ({
         </Space>
       </div>
       {/* {state.modalVisible && (
-         <PublishTemplate
-           saveTemplateServceId={saveTemplateServceId}
-           visible={state.modalVisible}
-           value={editorValue}
-           close={() => {
-             setState(draft => {
-               draft.modalVisible = false;
-             });
-           }}
-           fileType={'GREMLIN'}
-         />
-        )} */}
+          <PublishTemplate
+            saveTemplateServceId={saveTemplateServceId}
+            visible={state.modalVisible}
+            value={editorValue}
+            close={() => {
+              setState(draft => {
+                draft.modalVisible = false;
+              });
+            }}
+            fileType={'GREMLIN'}
+          />
+         )} */}
     </div>
   );
 };