+++ packages/gi-sdk/src/typing.ts
@@ -285,6 +285,11 @@ export type GIAssets = Partial<{
       [key: string]: any;
     };
   };
+  /** iconfont 图标 */
+  icons: {
+    icons: string[];
+    [keys: string]: any;
+  }[];
 }>;
 export interface LayoutConfig {
   // 支持的布局类型，默认为 force
@@ -330,6 +335,7 @@ export interface GINodeConfig {
   }[];
   logic: boolean;
   groupName: string;
+  default?: boolean;
 }
 export interface GIEdgeConfig {
   id: string;
@@ -347,6 +353,7 @@ export interface GIEdgeConfig {
   }[];
   logic: boolean;
   groupName: string;
+  default?: boolean;
 }
 
 export interface GIConfig {