+++ packages/gi-site/src/components/Navbar/SaveWorkbook.tsx
@@ -1,12 +1,13 @@
 import { SaveOutlined } from '@ant-design/icons';
+import { useShortcuts } from '@antv/gi-sdk';
 import { Graph } from '@antv/graphin';
 import { Button, notification, Tooltip } from 'antd';
 import * as React from 'react';
 import { useHistory } from 'react-router-dom';
+import $i18n from '../../i18n';
 import { useContext } from '../../pages/Analysis/hooks/useContext';
 import * as ProjectServices from '../../services/project';
 import type { IProject } from '../../services/typing';
-import $i18n from '../../i18n';
 import './index.less';
 
 interface SaveWorkbookProps {
@@ -119,6 +120,10 @@ const SaveWorkbook: React.FunctionComponent<SaveWorkbookProps> = props => {
       message: $i18n.get({ id: 'gi-site.components.Navbar.SaveWorkbook.SaveFailed', dm: '保存失败' }),
     });
   };
+
+  useShortcuts(['ctrl+s', 'command+s'], () => {
+    handleSave();
+  });
 
   return (
     <Tooltip title={$i18n.get({ id: 'gi-site.components.Navbar.SaveWorkbook.SaveCanvas', dm: '保存画布' })}>