+++ packages/gi-assets-tugraph-analytics/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @antv/gi-assets-tugraph-analytics
 
+## 0.2.4
+
+### Patch Changes
+
+- publish in 0720 pm
+- Updated dependencies
+  - @antv/gi-common-components@1.3.4
+  - @antv/gi-sdk@2.4.4
+
+## 0.2.3
+
+### Patch Changes
+
+- publish in 2023-07-20
+- Updated dependencies
+  - @antv/gi-common-components@1.3.3
+  - @antv/gi-sdk@2.4.3
+
+## 0.2.2
+
+### Patch Changes
+
+- weekly publish
+- Updated dependencies
+  - @antv/gi-common-components@1.3.2
+  - @antv/gi-sdk@2.4.2
+
 ## 0.2.1
 
 ### Patch Changes