+++ packages/gi-common-components/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @antv/gi-common-components
 
+## 1.3.4
+
+### Patch Changes
+
+- publish in 0720 pm
+
+## 1.3.3
+
+### Patch Changes
+
+- publish in 2023-07-20
+
+## 1.3.2
+
+### Patch Changes
+
+- weekly publish
+
 ## 1.3.1
 
 ### Patch Changes